helper functions move helper functions to own file change imports change use of functions to reflect
