+++ doc/hardware_notes.md
@@ -43,21 +43,21 @@ ubuntu@SeaHawk:~$
 
 * Sensors
 
-    * BME280 (unbridged)
-        * 0x77
+* BME280 (unbridged)
+      * 0x77
 
-    * BME280 (bridged)
-        * 0x76
+  * BME280 (bridged)
+    * 0x76
 
 * PCA9685 PWM Drivers
 
     Note that address 0x70 is the "all call" address for the controller chip on the HAT. All boards will respond to address 0x70 - regardless of the address jumper settings.
 
-    * Servo Driver Hat
-        * 0x40
+  * Servo Driver Hat
+    * 0x40
 
-    * Motor Driver Hat
-        * 0x60
+  * Motor Driver Hat
+    * 0x60
 
-    * Servo Driver (A0 Bridged)
-        * 0x41
+  * Servo Driver (A0 Bridged)
+    * 0x41
