+++ src/seahawk_rov/seahawk_rov/logic_tube_servo.py
@@ -32,21 +32,8 @@ from std_msgs.msg import Float32
 # adafruit servokit for hat
 from adafruit_servokit import ServoKit
 
-
-# linear interpolation helper function
-def lerp(old_min:float, old_max:float, new_min:int, new_max:int, old_value:float):
-    '''linear interpolate helper function'''
-    old_range = old_max - old_min
-    new_range = new_max - new_min
-    new_value = (((old_value - old_min) * new_range) / old_range) + new_min
-    return new_value
-
-
-# clamp helper function
-# helps keep us from starting fires
-def clamp(num, minimum, maximum):
-  '''clamp helper function'''
-  return max(min(minimum, num), maximum)
+# import helper functions
+import seahawk_rov
 
 class LogicTubeServo:
     def __init__(self, node, i2c):
@@ -66,4 +53,4 @@ class LogicTubeServo:
 
 
     def receive_drive_camera(self, message):
-        self.kit.servo[self.drive_cam_servo].angle = int(lerp(-1.0, 1.0, 0, 3000, clamp(message.data, -1, 1)))
+        self.kit.servo[self.drive_cam_servo].angle = int(seahawk_rov.lerp(-1.0, 1.0, 0, 3000, seahawk_rov.clamp(message.data, -1, 1)))
